vest_plus::combinators::dependent

Trait DepCombinator

pub trait DepCombinator {
    type Key;
    type Val;
    type Body: Consistency<Val = Self::Val>;

    // Required methods
    spec fn apply(&self, key: Self::Key) -> Self::Body;
    spec fn recover(&self, value: Self::Val) -> Self::Key;
    proof fn lemma_recover_consistent(&self, key: Self::Key, value: Self::Val);
}

A family of dependent combinators indexed by a key type.

Used to constrain the Tail combinator in a Bind<Head, Tail>.

While possible, most users will use the provided dependent combinators (TVOr, TLVal, VariedLen, etc.) rather than implementing this trait directly.

Required Associated Types

type Key

The type of keys parsed by the head combinator/to be recovered during serialization.

type Val

The type of values consumed/produced by the body combinator.

type Body: Consistency<Val = Self::Val>

The type of the body combinator produced by apply.

Required Methods

spec fn apply(&self, key: Self::Key) -> Self::Body

Given a key, produce the body combinator for that key.

spec fn recover(&self, value: Self::Val) -> Self::Key

Given a body value, recover the key used to produce the body combinator.

proof fn lemma_recover_consistent(&self, key: Self::Key, value: Self::Val)

ensures
self.apply(key).consistent(value) ==> self.recover(value) == key,

Recover must agree with any key that makes the body’s value consistent.