vest_plus::combinators::recursive

Module spec

Specification trait implementations for this combinator.

Structs

NestedBracesBody

SpecRecBody for the nested-braces example.

NestedBracesMapper

Mapper between Sum<NestedBracesT, ()> and NestedBracesT.

Enums

NestedBracesT

Example recursive value type: nested braces {...} or empty \0.

Traits

EquivSerializersGeneralRecBody

Serializer equivalence invariant preservation for recursive bodies.

GoodSerializerRecBody

Serializer’s properties preservation for recursive bodies.

NoLookAheadRecBody

No-lookahead invariant preservation for recursive bodies.

NonMalleableRecBody

Non-malleability preservation for recursive bodies.

NonTailFmtRecBody

DPS serializer’s properties preservation for recursive bodies.

SPRoundTripDpsRecBody

DPS serialize-parse roundtrip preservation for recursive bodies. Similar to Star, the body must also be NonTailFmt.

SoundParserRecBody

Soundness preservation for recursive bodies.

SpecRecBody

Defines one level of a recursive format for use with super::Fix.

Functions

good_serializer_fn

nested_braces_body

no_lookahead_parser

non_malleable_parser

non_tail_fmt_dps

parser_pair_some

parser_specs

serializer_specs

sound_parser

sp_roundtrip_dps_parser

Type Aliases

BundledSpecs

ParserFnSpecs

ParserSpecs

Bundled triple of parser, consistency, and byte-length spec functions.

SerializerFnSpecs

SerializerSpecs

A bundled non-DPS serializer: pairs a SerializerFnSpec with a ByteLenFnSpec.