vest_plus::combinators::recursive::spec

Trait RecBody

pub trait RecBody {
    type Val;

    // Required methods
    spec fn parse_body(
        &self,
        rec: ParserFnSpec<Self::Val>,
    ) -> ParserFnSpec<Self::Val>;
    spec fn byte_len_body(
        &self,
        rec: ByteLenFnSpec<Self::Val>,
    ) -> ByteLenFnSpec<Self::Val>;
    spec fn consistent_body(
        &self,
        rec: PredFnSpec<Self::Val>,
    ) -> PredFnSpec<Self::Val>;
    proof fn lemma_body_preservation(
        &self,
        p_rec: ParserFnSpec<Self::Val>,
        c_rec: PredFnSpec<Self::Val>,
        b_rec: ByteLenFnSpec<Self::Val>,
    );
}

Defines one level of a recursive format for use with super::Fix.

Required Associated Types

type Val

The type of values parsed/serialized by this recursive format.

Required Methods

spec fn parse_body(&self, rec: ParserFnSpec<Self::Val>) -> ParserFnSpec<Self::Val>

Parser for one level, given a recursive callback.

spec fn byte_len_body( &self, rec: ByteLenFnSpec<Self::Val>, ) -> ByteLenFnSpec<Self::Val>

Byte-length function for one level, given a recursive callback.

spec fn consistent_body(&self, rec: PredFnSpec<Self::Val>) -> PredFnSpec<Self::Val>

Consistency predicate for one level, given a recursive callback.

proof fn lemma_body_preservation( &self, p_rec: ParserFnSpec<Self::Val>, c_rec: PredFnSpec<Self::Val>, b_rec: ByteLenFnSpec<Self::Val>, )

requires
good_parser(p_rec, c_rec, b_rec),
ensures
good_parser(
    self.parse_body(p_rec),
    self.consistent_body(c_rec),
    self.byte_len_body(b_rec),
),

good_parser is preserved through one level of unfolding.