vest_plus::combinators::recursive::spec

Trait SpecRecBody

pub trait SpecRecBody {
    type T;
    type Body: SpecCombinator<T = Self::T>;

    // Required method
    spec fn spec_body(&self, rec: BundledSpecs<Self::T>) -> Self::Body;
}

Defines one level of a recursive format for use with super::Fix.

Required Associated Types

type T

type Body: SpecCombinator<T = Self::T>

Required Methods

spec fn spec_body(&self, rec: BundledSpecs<Self::T>) -> Self::Body

Define one recursive unfolding, where rec provides callbacks for all recursive positions. the callback for recursive positions in the body.