vest_plus::combinators::recursive

Struct Fix

pub struct Fix<const LIMIT: usize, Body>(pub Body);

Bounded fixpoint combinator: recursive format up to LIMIT levels of nesting.

Parsing semantics: parses with Body, which may recursively refer to Fix (up to LIMIT times).

TODO: this combinator currently only supports parsing; we should add serialization support as well.

Tuple Fields
0: Body
Implementations

impl<const LIMIT: usize, Body: RecBody> Fix<LIMIT, Body>

pub open spec fn spec_parse_with_gas( &self, gas: nat, input: Seq<u8>, ) -> Option<(int, Body::Val)>

{ self.0.parse_body(self.parse_callback(gas))(input) }

Parse with a given amount of gas. Unfolds one level via body.parse_body with a callback that recurses at gas - 1.

pub open spec fn consistent_with_gas(&self, gas: nat, v: Body::Val) -> bool

{ self.0.consistent_body(self.consistent_callback(gas))(v) }

Consistency check with given gas.

pub open spec fn byte_len_with_gas(&self, gas: nat, v: Body::Val) -> nat

{ self.0.byte_len_body(self.byte_len_callback(gas))(v) }

Byte-length computation with given gas.

pub open spec fn parse_callback(&self, gas: nat) -> ParserFnSpec<Body::Val>

{
    |ibuf: Seq<u8>| {
        if gas > 0 { self.spec_parse_with_gas((gas - 1) as nat, ibuf) } else { None }
    }
}

Recursive parser callback: recurses at gas - 1, returns None at zero.

pub open spec fn consistent_callback(&self, gas: nat) -> PredFnSpec<Body::Val>

{
    |vv: Body::Val| {
        if gas > 0 { self.consistent_with_gas((gas - 1) as nat, vv) } else { false }
    }
}

Recursive consistency callback: recurses at gas - 1, returns false at zero.

pub open spec fn byte_len_callback(&self, gas: nat) -> ByteLenFnSpec<Body::Val>

{
    |vv: Body::Val| {
        if gas > 0 { self.byte_len_with_gas((gas - 1) as nat, vv) } else { 0 }
    }
}

Recursive byte-length callback: recurses at gas - 1, returns 0 at zero.