vest_plus::combinators::tail

Struct Eof

pub struct Eof;

End-of-file combinator: denotes the “EOF”.

Parsing semantics: succeeds only if the input is empty, producing ().

Implements AdmitsUniqueVal.

Note

The DPS serialization always replaces the output buffer with the empty sequence, so Eof should only appear at the end of a format (and the trait system enforces this).