vest_plus::core::spec

Trait Serializability

pub trait Serializability: SpecSerializerDps {
    // Provided method
    open spec fn serializable(&self, v: Self::ST, obuf: Seq<u8>) -> bool { ... }
}

DPS-specific serializability constraints.

Some combinators impose conditions on which (value, obuf) pairs are valid for DPS serialization. For example, crate::combinators::Opt requires the inner parser to fail on obuf when the value is None; crate::combinators::Star requires the inner parser to fail on obuf for termination; crate::combinators::Choice requires the left parser to fail when serializing the right branch.

Most primitive combinators leave this as true.

Provided Methods

open spec fn serializable(&self, v: Self::ST, obuf: Seq<u8>) -> bool

{ true }

Returns true if v can be serialized with output buffer obuf.