vest_plus::core::spec

Trait SpecParser

pub trait SpecParser {
    type PVal;

    // Required method
    spec fn spec_parse(&self, ibuf: Seq<u8>) -> Option<(int, Self::PVal)>;
}

Parser specification.

Required Associated Types

type PVal

The type of parsed values.

Required Methods

spec fn spec_parse(&self, ibuf: Seq<u8>) -> Option<(int, Self::PVal)>

Attempts to parse a value from ibuf.

Returns Some((n, v)) on success, where n bytes were consumed and v is the parsed value, or None on failure.